renamed original calpost_ncf as cf, and made ncf2 to ncf
